requirements_test.txt
requests==2.25.1
pytest==6.2.4
pytest-homeassistant-custom-component==0.4.1
